requirements.txt
backcall
bleach
colorama
decorator
defusedxml
Django
django-extensions
django-filter
djangorestframework
entrypoints
ipykernel
ipython-genutils
jedi
Jinja2
jsonschema
jupyter-client
jupyter-core
Markdown
MarkupSafe
mistune
nbconvert
nbformat
notebook
pandocfilters
parso
pickleshare
prometheus-client
prompt-toolkit
Pygments
python-dateutil
pytz
pywinpty
pyzmq
Send2Trash
six
terminado
testpath
tornado
traitlets
wcwidth
webencodings
